organize files and add poses list
