+++ src/Components/PosesList.js
@@ -1,0 +1,33 @@
+// TODO: write code that makes a request to yoga API and YouTube API and renders a list of the names of all the poses as links that return the user to the Pose route for the pose they click on
+
+// TODO: Add function to request YouTube video
+
+import axios from "axios";
+import { useState, useEffect } from "react";
+
+export function PosesList() {
+
+    const [allPoses, setAllPoses] = useState([]);
+
+    useEffect(() => {
+        async function getPose() {
+            const res = await axios.get("https://lightning-yoga-api.herokuapp.com/yoga_poses")
+            .then((res) => {
+                setAllPoses(res.data.items);
+            })
+            .catch((error) => {
+                console.error(error)
+            })           
+        }
+        getPose()
+    }, []);
+
+    return (
+        <>
+        <h1>Choose a Pose</h1>
+        {allPoses ? 
+        allPoses.map((pose, i) => <p key={i}>{pose.sanskrit_name}</p>)            
+        : "Loading..."}
+        </>
+    )
+}
