--- src/PosesList.js
@@ -1,1 +1,0 @@
-// TODO: write code that makes a request to yoga API and YouTube API and renders a list of the names of all the poses as links that return the user to the Pose route for the pose they click on`
